backend/requirements.txt
aiobotocore==2.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
async-timeout==5.0.1
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
billiard==4.2.1
boto3==1.38.27
botocore==1.38.27
build==1.2.2.post1
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.13
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
cryptography==45.0.5
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.23.1
durationpy==0.10
dvc==3.60.1
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.18.0
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.2.0
frozenlist==1.7.0
fsspec==2025.5.1
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
googleapis-common-protos==1.70.0
grandalf==0.8
grpcio==1.73.1
gto==1.7.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
iterative-telemetry==0.0.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
kubernetes==33.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.22.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
posthog==6.0.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.0
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.1
s3fs==2025.5.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
scmrepo==3.3.11
semver==3.0.4
sentence-transformers==5.0.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.2
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
transformers==4.53.0
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
voluptuous==0.15.2
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zc.lockfile==3.0.post1
zipp==3.23.0
geopy>=2.0

python-dotenv
requests
